Don't consider artifacts when deciding if a chest is empty. Closes #84
Previously, we looked for any items in a chest before deciding whether
to refill it. This included artifacts, which get generated for every
quest of that type, and stay there until a party picks up the quest and
completes it. This resulted in most chests never getting filled. So we
fill chests even if they have artifacts now.